% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicontacts-ui.R
\name{epicontacts_ui}
\alias{epicontacts_ui}
\title{Web interface for epicontacts}
\usage{
epicontacts_ui()
}
\description{
This function opens up an application in a web browser for an interactive use
of RECON's epicontacts package.
}
\seealso{
The epicontacts package: \url{http://www.repidemicsconsortium.org/epicontacts/}
}
